pyproject.toml
[tool.poetry]
name = "yamlimg"
version = "1.0.0"
description = "A way to store images in YAML."
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"
repository = "https://github.com/vcokltfre/yamlimg"
keywords = ["yaml", "image", "storage"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
typer = "^0.4.0"
Pillow = "^9.0.1"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pre-commit = "^2.17.0"
taskipy = "^1.10.1"

[tool.poetry.scripts]
yamlimg = "yamlimg.cli:app"

[tool.taskipy.tasks]
lint = "black . && isort --profile black ."
pre-commit = "pre-commit install"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
